backend/pyproject.toml
[project]
name = "odoo-dev-dashboard"
version = "1.0.0"
description = "Odoo Dev Dashboard — Django + Vue.js rewrite"
requires-python = ">=3.12"
dependencies = [
    "django>=5.0",
    "djangorestframework>=3.15",
    "channels>=4.0",
    "channels-redis>=4.0",
    "daphne>=4.0",
    "docker>=7.0",
    "psycopg2-binary>=2.9",
    "jupyter_client>=8.0",
    "ipykernel>=6.0",
    "apscheduler>=3.10",
    "python-dotenv>=1.0",
    "django-cors-headers>=4.0",
    "psutil>=5.9",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=8.0",
    "pytest-django>=4.8",
]

[tool.hatch.build.targets.wheel]
packages = ["core", "apps", "services"]
